+++ src/assets/icons/MessageRoundIcon.tsx
@@ -2,8 +2,9 @@ import React from 'react';
 
 import {Svg, Path, Circle} from 'react-native-svg';
 
+import {palette} from '@theme';
+
 import {IconBase} from '../../components/Icon/Icon';
-import {palette} from '../../theme/theme';
 
 export function MessageRoundIcon({
   size = 48,
