+++ src/theme/index.ts
@@ -1,1 +1,2 @@
 export * from './theme';
+export * from './colors';
